chore: render code style clean and update (#2477)
