+++ src/core/render/index.js
@@ -82,7 +82,7 @@ export function Render(Base) {
         }
       }
 
-      this._renderTo(markdownElm, html);
+      dom.setHTML(markdownElm, html);
 
       // Execute markdown <script>
       if (
@@ -274,13 +274,6 @@ export function Render(Base) {
       }
     }
 
-    _renderTo(el, content, replace) {
-      const node = dom.getNode(el);
-      if (node) {
-        node[replace ? 'outerHTML' : 'innerHTML'] = content;
-      }
-    }
-
     _renderSidebar(text) {
       const { maxLevel, subMaxLevel, loadSidebar, hideSidebar } = this.config;
       const sidebarEl = dom.getNode('aside.sidebar');
@@ -294,7 +287,7 @@ export function Render(Base) {
         return null;
       }
 
-      this._renderTo('.sidebar-nav', this.compiler.sidebar(text, maxLevel));
+      dom.setHTML('.sidebar-nav', this.compiler.sidebar(text, maxLevel));
 
       sidebarToggleEl.setAttribute('aria-expanded', !isMobile());
 
@@ -366,7 +359,7 @@ export function Render(Base) {
       const html = this.compiler.compile(text);
 
       ['.app-nav', '.app-nav-merged'].forEach(selector => {
-        this._renderTo(selector, html);
+        dom.setHTML(selector, html);
         this.#addTextAsTitleAttribute(`${selector} a`);
       });
     }
@@ -507,7 +500,7 @@ export function Render(Base) {
         rootElm.style.setProperty('--cover-bg', mdCoverBg);
       }
 
-      this._renderTo('.cover-main', html);
+      dom.setHTML('.cover-main', html);
 
       // Button styles
       dom
@@ -563,7 +556,7 @@ export function Render(Base) {
         html += tpl.main(config);
 
         // Render main app
-        this._renderTo(el, html, true);
+        dom.setHTML(el, html, true);
       } else {
         this.rendered = true;
       }
